pyproject.toml
[project]
name = "rg-agno-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain>=0.3.27",
    "mcp>=1.14.0",
    "openai>=1.107.3",
    "sqlalchemy[asyncio]>=2.0.43",
    "ruamel-yaml>=0.18.6,<0.19.0", # YAML处理
    "cachetools==5.3.3", # 缓存工具
    "filelock==3.15.4", # 文件锁
    "itsdangerous==2.1.2", # 安全签名，用于 SessionMiddleware
    "langchain-openai>=0.3.33",
    "langchainhub>=0.1.21",
    "httpx-sse>=0.4.1",
    "sqlmodel>=0.0.25",
    "aiomysql>=0.2.0",
]
[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
